+++ src/cmd_read.cpp
@@ -50,12 +50,7 @@ static bool readLine(ARState &state, int fd, const ArrayObject &argvObj, unsigne
   for (bool prevIsBackslash = false;;
        prevIsBackslash = backslash && ch == '\\' && !prevIsBackslash) {
     while (true) {
-      const ReadWithTimeoutParam param = {
-          .retry = false,
-          .timeoutMSec = timeoutMSec,
-          .interruptSet = nullptr,
-      };
-      readSize = readWithTimeout(fd, &ch, 1, param);
+      readSize = readWithTimeout(fd, &ch, 1, {.retry = false, .timeoutMSec = timeoutMSec});
       if (readSize < 0) {
         if (errno == EAGAIN) {
           continue;