+++ src/keycode.h
@@ -38,7 +38,6 @@ inline bool isCaretTarget(int ch) { return (ch >= '@' && ch <= '_') || ch == '?'
 struct ReadWithTimeoutParam {
   bool retry;
   int timeoutMSec;
-  const sigset_t *interruptSet; // if not null, only interrupt by specified signals
 };
 
 /**
@@ -52,15 +51,10 @@ struct ReadWithTimeoutParam {
  * if has error, return -1 and set errno
  * otherwise, return non-negative number
  */
-ssize_t readWithTimeout(int fd, char *buf, size_t bufSize, const ReadWithTimeoutParam &param);
+ssize_t readWithTimeout(int fd, char *buf, size_t bufSize, ReadWithTimeoutParam param);
 
 inline ssize_t readRetryWithTimeout(int fd, char *buf, size_t bufSize, int timeoutMSec) {
-  const ReadWithTimeoutParam param = {
-      .retry = true,
-      .timeoutMSec = timeoutMSec,
-      .interruptSet = nullptr,
-  };
-  return readWithTimeout(fd, buf, bufSize, param);
+  return readWithTimeout(fd, buf, bufSize, {.retry = true, .timeoutMSec = timeoutMSec});
 }
 
 class KeyCodeReader {