+++ test/interactive/interactive_test3.cpp
@@ -447,13 +447,13 @@ TEST_F(InteractiveTest, changeTCPGRPInChild) {
   this->invoke("--quiet", "--norc");
 
   ASSERT_NO_FATAL_FAILURE(this->expect(PROMPT));
-  std::string out = ": Job = %1\n" + PROMPT;
-  std::string err = format("[warn] retry readLine, caused by `%s'\n", strerror(EIO));
   {
     auto cleanup = this->withTimeout(400);
     ASSERT_NO_FATAL_FAILURE(this->sendLineAndExpect(
-        "call $BIN_NAME -c 'shctl set monitor; ls > /dev/null;' &", out.c_str(), err.c_str()));
+        "call $BIN_NAME -c 'shctl set monitor; ls > /dev/null;' &", ": Job = %1"));
   }
+  std::string err = format("[warn] retry readLine, caused by `%s'\n", strerror(EIO));
+  ASSERT_NO_FATAL_FAILURE(this->sendLineAndExpect("", "", err.c_str()));
   ASSERT_NO_FATAL_FAILURE(this->sendLineAndWait("exit", 0));
 }