+++ CHANGELOG.md
@@ -36,6 +36,7 @@
     - now only replace first occurrence of pattern
 - add ``start``, ``end`` method to ``RegexMatch`` type. now get offsets of match string
 - improve error message of builtin ``shctl`` command. now show subcommand name
+- in ``LineEditor#readLine``, automatically refresh line when receive ``SIGWINCH``
 
 #### LSP